requirements.txt
aiofiles==24.1.0
aiohttp==3.12.15
aiosqlite==0.21.0
coloredlogs==15.0.1
Jinja2==3.1.6
pyTelegramBotAPI==4.28.0
PyYAML==6.0.2